Modified the generated script to exit on errors
Previous output example:

#############################################
# Generating Swift clients from OpenAPI! :) #
#############################################

generate-api.sh: line 13: openapi-generator: command not found
cp: ./output/OpenAPIClient/Classes/OpenAPIs/: No such file or directory

#########################################
# Swift client created successfully! :) #
#########################################

As you can see, the script failed but it says it created successfully. I want to show an error if there is an error like the one provided and only show the success messages if things go well

+++ Sources/ReccoHeadless/Api/generate-api.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Exit immediately if a command exits with a non-zero status.
+set -e
+
 echo ""
 echo "#############################################"
 echo "# Generating Swift clients from OpenAPI! :) #"
@@ -10,24 +13,30 @@ echo ""
 rm -rf ./Generated
 
 # Generate new Swift client
-openapi-generator generate \
-	-i ./sf-backend-open-api.json \
-	-g swift5 \
-	--model-name-suffix DTO \
+if openapi-generator generate \
+    -i ./sf-backend-open-api.json \
+    -g swift5 \
+    --model-name-suffix DTO \
     --additional-properties responseAs=AsyncAwait,nonPublicApi=true \
     --type-mappings=date-time=Date \
     --type-mappings=date=String \
-	-o ./output
+    -o ./output; then
+    # Move new client files to code-gen
+    cp -R ./output/OpenAPIClient/Classes/OpenAPIs/ ./Generated
 
+    # Remove output folder
+    rm -rf ./output
 
-# Move new client files to code-gen
-cp -R ./output/OpenAPIClient/Classes/OpenAPIs/ ./Generated
-
-# Remove output folder
-rm -rf ./output
-
-echo ""
-echo "#########################################"
-echo "# Swift client created successfully! :) #"
-echo "#########################################"
-echo ""
+    echo ""
+    echo "#########################################"
+    echo "# Swift client created successfully! :) #"
+    echo "#########################################"
+    echo ""
+else
+    echo ""
+    echo "#########################################"
+    echo "# Error: Swift client generation failed! #"
+    echo "#########################################"
+    echo ""
+    exit 1
+fi
